requirements.txt
gevent==24.11.1
greenlet==3.1.1
PyQt5==5.15.11
PyQt5-Qt5==5.15.15
PyQt5_sip==12.15.0
python-dotenv==1.0.1
setuptools==75.5.0
websocket-client==1.8.0
zope.event==5.0
zope.interface==7.1.1
